fix: move useMatchingUsers to hooks file
